Comments about the project
I really like how your website looks! Its descriptive enough about you and all requirements of the rubric are fulfilled. You can mention in the README about the use of GenAI if any, and how you did it, and what page you made with a GenAI tool. But all in all, amazing work! And I love the drawer thing you implemented in the About page.